+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Automatic entry nodes list downloading
+- IRI healthchecks on startup without throwing an error.
+- Actively remove peers, if the limit is trespassed at any point for any reason.
 
 ## [0.1.11] - 2017-12-16